Tesis/requirements.txt
absl-py==2.1.0
attrs==24.2.0
blinker==1.8.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cryptography==46.0.5
cycler==0.12.1
decorator==5.1.1
filelock==3.24.3
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2026.2.0
idna==3.11
imageio==2.36.1
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
jax==0.4.35
jaxlib==0.4.35
Jinja2==3.1.4
kiwisolver==1.4.7
lap==0.5.12
MarkupSafe==3.0.2
matplotlib==3.9.2
mediapipe==0.10.14
ml_dtypes==0.5.0
moviepy==2.1.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.2
opencv-contrib-python==4.10.0.84
opencv-python==4.13.0.92
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
packaging==24.1
pillow==10.4.0
polars==1.38.1
polars-runtime-32==1.38.1
proglog==0.1.10
protobuf==4.25.5
psutil==7.2.2
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.8.0
pyodbc==5.2.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.3
requests==2.32.5
scipy==1.14.1
six==1.16.0
sounddevice==0.5.1
sympy==1.14.0
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.1
typing_extensions==4.15.0
ultralytics==8.4.14
ultralytics-thop==2.0.18
urllib3==2.6.3
Werkzeug==3.1.0
